Moved HeadComman from the root of the lobbyheads package to `head` package
